Рефакторинг модуля users
