+++ project/apps/users/src/app/auth/auth.controller.ts
@@ -1,11 +1,7 @@
-import { Post, Body, Get, Param, HttpCode, HttpStatus, Controller, UseGuards, Req } from '@nestjs/common';
+import { Post, HttpCode, HttpStatus, Controller, Req, UseGuards } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import { AuthService } from './auth.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { fillObject } from '@project/util/util-core';
-import { UserRdo } from './rdo/user.rdo';
 import { LoggedUserRdo } from './rdo/logged-user.rdo';
-import { MongoidValidationPipe } from '@project/shared/shared-pipes';
 import { JwtAuthGuard } from './guards/jwt-auth.guard';
 import { NotifyService } from '../notify/notify.service';
 import { LocalAuthGuard } from './guards/local-auth-guard';
@@ -19,17 +15,6 @@ export class AuthController {
     private readonly authService: AuthService,
     private readonly notifyService: NotifyService,
   ) {}
-  @ApiResponse({
-    status: HttpStatus.CREATED,
-    description: 'The new user has been successfully created.'
-  })
-  @Post('register')
-  public async create(@Body() dto: CreateUserDto) {
-    const newUser = await this.authService.register(dto);
-    const { email, name } = newUser;
-    await this.notifyService.registerSubscriber({ email, name })
-    return fillObject(UserRdo, newUser);
-  }
 
   @UseGuards(LocalAuthGuard)
   @ApiResponse({
@@ -47,18 +32,6 @@ export class AuthController {
     return this.authService.createUserToken(user);
   }
 
-  @ApiResponse({
-    type: UserRdo,
-    status: HttpStatus.OK,
-    description: 'User found'
-  })
-  @UseGuards(JwtAuthGuard)
-  @Get(':id')
-  public async show(@Param('id', MongoidValidationPipe) id: string) {
-    const existUser = await this.authService.getUser(id);
-    return fillObject(UserRdo, existUser);
-  }
-
   @UseGuards(JwtRefreshGuard)
   @Post('refresh')
   @HttpCode(HttpStatus.OK)
@@ -69,7 +42,6 @@ export class AuthController {
   public async refreshToken(@Req() { user }: RequestWithUser) {
     return this.authService.createUserToken(user);
   }
-
   @UseGuards(JwtAuthGuard)
   @Post('check')
   public async checkToken(@Req() { user: payload }: RequestWithTokenPayload) {
