+++ project/apps/users/src/app/auth/auth.http
@@ -1,21 +1,3 @@
-# Create a new user
-POST http://localhost:3001/api/auth/register HTTP/1.1
-Content-Type: application/json
-
-{
-  "email": "user@notfound.ru",
-  "name": "Keks",
-  "password": "123456"
-}
-
-###
-
-# Get user by ID
-GET http://localhost:3001/api/auth/6582d13b839eca516d695a99 HTTP/1.1
-Authorization: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGU2ZWRlZmVmN2VmNWEwNWIzZDY4NjMiLCJlbWFpbCI6InVzZXJAbm90Zm91bmQubG9jYWwiLCJuYW1lIjoiS2VrcyIsImF2YXRhciI6IiIsImlhdCI6MTY5MzA1ODUxNCwiZXhwIjoxNjkzMDU5MTE0fQ.FLY1FC0tF0OFHKNA2yF6bnPNzK1RZH8cKzlDmh-xz40
-
-###
-
 # Login user by email and password
 POST http://localhost:3001/api/auth/login HTTP/1.1
 Content-Type: application/json
