+++ project/apps/users/src/app/auth/auth.service.ts
@@ -1,9 +1,8 @@
-import { ConflictException, Injectable, NotFoundException, UnauthorizedException, Inject } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException, Inject } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService, ConfigType } from '@nestjs/config';
 import { BlogUserRepository } from '../blog-user/blog-user.repository';
 import { BlogUserEntity } from '../blog-user/blog-user.entity';
-import { CreateUserDto } from './dto/create-user.dto';
 import { LoginUserDto } from './dto/login-user.dto';
 import { AuthError } from './auth.constant';
 import { User } from '@project/shared/app-types';
@@ -22,55 +21,23 @@ export class AuthService {
     private readonly refreshTokenService: RefreshTokenService,
   ) {}
 
-  public async register(dto: CreateUserDto) {
-    const {email, name, password} = dto;
-
-    const blogUser = {
-      email,
-      name,
-      avatar: '',
-      passwordHash: ''
-    };
-
-    const existUser = await this.blogUserRepository
-      .findByEmail(email);
-
-    if (existUser) {
-      throw new ConflictException(AuthError.AUTH_USER_EXISTS);
-    }
-
-    const userEntity = await new BlogUserEntity(blogUser)
-      .setPassword(password);
-
-    return this.blogUserRepository
-      .create(userEntity);
-  }
-
   public async verifyUser(dto: LoginUserDto) {
     const {email, password} = dto;
     const existUser = await this.blogUserRepository.findByEmail(email);
-
     if (!existUser) {
       throw new NotFoundException(AuthError.AUTH_USER_NOT_FOUND);
     }
-
     const blogUserEntity = new BlogUserEntity(existUser);
     if (!await blogUserEntity.comparePassword(password)) {
       throw new UnauthorizedException(AuthError.AUTH_USER_PASSWORD_WRONG);
     }
-
     return blogUserEntity.toObject();
-  }
-
-  public async getUser(id: string) {
-    return this.blogUserRepository.findById(id);
   }
 
   public async createUserToken(user: User) {
     const accessTokenPayload = createJWTPayload(user);
     const refreshTokenPayload = { ...accessTokenPayload, tokenId: crypto.randomUUID() };
     await this.refreshTokenService.createRefreshSession(refreshTokenPayload)
-
     return {
       accessToken: await this.jwtService.signAsync(accessTokenPayload),
       refreshToken: await this.jwtService.signAsync(refreshTokenPayload, {
