+++ project/apps/users/src/app/blog-user/blog-user.constant.ts
@@ -1,1 +1,3 @@
 export const SALT_ROUNDS = 10;
+export const AUTH_USER_EMAIL_NOT_VALID = 'The email is not valid';
+export const AUTH_USER_EXISTS = 'User with this email exists';
