+++ project/apps/users/src/app/blog-user/blog-user.http
@@ -1,0 +1,17 @@
+# Create a new user
+POST http://localhost:3001/api/user/register HTTP/1.1
+Content-Type: application/json
+
+{
+  "email": "user2@notfound.local",
+  "name": "Keks",
+  "password": "123456"
+}
+
+###
+
+# Get user by ID
+GET http://localhost:3001/api/user/64eb1477caa87cd6c8483985 HTTP/1.1
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGViMTQ3N2NhYTg3Y2Q2Yzg0ODM5ODUiLCJlbWFpbCI6InVzZXIyQG5vdGZvdW5kLmxvY2FsIiwibmFtZSI6Iktla3MiLCJhdmF0YXIiOiIiLCJpYXQiOjE2OTMxMjc5MDIsImV4cCI6MTY5MzEyODUwMn0.29bYEnhGvq66Hnpi3ttlIb9G01xBagwwKprlVMAZH74
+
+###
